Fix conftest reliance
