+++ failed_tests.txt
@@ -1,1 +1,1 @@
-("test_fail2", "failed")
+('test_fail', 'failed')
