+++ src/pytest_focus.py
@@ -2,9 +2,54 @@
 """
 Plug in for pytest that sends push notifications for failed tests
 """
-
+import os
+import sys
+from sys import platform
+import webbrowser
 import pytest
 from _pytest.terminal import TerminalReporter
+
+
+def mac_notify(title, subtitle, message):
+    """ Handles Mac Notification """
+    push_t = "-title {!r}".format(title)
+    push_s = "-subtitle {!r}".format(subtitle)
+    push_m = "-message {!r}".format(message)
+    os.system(
+        'terminal-notifier {} -activate "com.apple.Terminal"'.format(
+            " ".join([push_m, push_t, push_s])
+        )
+    )
+    sys.stdout.write("It appears you have failing test cases...\n")
+
+
+def win_notify(title, message):
+    """ Handles windows notifications """
+    os.system("notify-send -i error {}".format(" ".join([title, message])))
+
+
+def linux_notify(title, message):
+    """ Handles linux notifications """
+    os.system("notify-send {} -u critical".format(" ".join([title, message])))
+
+
+# The notifier function
+def notify(title, message, subtitle="Test Case Failed."):
+    """ Send a notification to the user's screen """
+    if platform == "darwin":
+        mac_notify(title, subtitle, message)
+    elif platform in ("linux", "linux2"):
+        linux_notify(title, message)
+    elif platform == "win32":
+        win_notify(title, message)
+
+
+def todo_list(test_details):
+    """ List of failed test cases in txt file """
+    # pylint: disable=invalid-name
+    f = open("failed_tests.txt", "w+")
+    f.write(test_details)
+    webbrowser.open("failed_tests.txt")
 
 
 def pytest_addoption(parser):
@@ -50,15 +95,20 @@ class FocusingTerminalReporter(TerminalReporter):
         if report.failed:
             if self.isatty:
                 self.rewrite("")
-            # self.print_failure(report)
+            self.print_failure(report)
 
     def pytest_runtest_logreport(self, report):
         """ Shows failures and errors as tests are running """
         TerminalReporter.pytest_runtest_logreport(self, report)
         if report.failed and not hasattr(report, "wasxfail"):
+            notify(str(getattr(report, "head_line")), str(getattr(report, "outcome")))
+            test_details = str(
+                (str(getattr(report, "head_line")), str(getattr(report, "outcome")))
+            )
+            todo_list(test_details)
             if self.verbosity <= 0:
                 self._tw.line()
-            # self.print_failure(report)
+            self.print_failure(report)
 
     def print_failure(self, report):
         """ sends push notifications as test cases fail """
@@ -76,22 +126,5 @@ class FocusingTerminalReporter(TerminalReporter):
                 elif when == "teardown":
                     msg = "ERROR at teardown of " + msg
                 self.write_sep("_", msg)
-                # subprocess.call("notify-send " + msg)
                 if not self.config.getvalue("usepdb"):
                     self._outrep_summary(report)
-
-
-# def pytest_test():
-#     """
-#     the plug-in for pytest
-#     """
-#     os_name = "unknown"
-#
-#     if platform in ("linux", "linux2"):
-#         os_name = "linux"  # linux
-#     elif platform == "darwin":
-#         os_name = "Mac"  # OS X
-#     elif platform == "win32":
-#         os_name = "windows"  # Windows...
-#
-#     return os_name
