+++ tests/test_failure.py
@@ -5,7 +5,7 @@ Test cases designed to see if plugin works
 import pytest
 
 
-@pytest.mark.xfail
+# @pytest.mark.xfail
 def test_fail():
     """ Failing test """
